// src/models/IMovie.ts
export interface IMovie {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  genres?: { id: number; name: string }[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: Date;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}